+++ src/app/shared/services/question-control.service.ts
@@ -1,7 +1,7 @@
 import { Injectable }   from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 
-import { QuestionBase } from '../models/question.model';
+import { QuestionBase } from '../question-base';
 
 @Injectable()
 export class QuestionControlService {
